+++ src/rules.ts
@@ -27,3 +27,5 @@ import './polymer/databind-with-unknown-property';
 import './polymer/element-before-dom-module';
 import './polymer/set-unknown-attribute';
 import './polymer/unbalanced-delimiters';
+
+import './polymer/elements/rules';